+++ slice_utils.go
@@ -1,0 +1,11 @@
+package ezutil
+
+func MapSlice[T any, U any](input []T, mapperFunc func(T) U) []U {
+	output := make([]U, len(input))
+
+	for i, v := range input {
+		output[i] = mapperFunc(v)
+	}
+
+	return output
+}
